feat: add GPX parser, point container, and Multiple task manager modules

+++ Software/X-Track/USER/App/Common/DataProc/DataProc_Def.h
@@ -115,8 +115,8 @@ typedef enum
 typedef struct
 {
     SysConfig_Cmd_t cmd;
-    float longitude;
-    float latitude;
+    double longitude;
+    double latitude;
     int16_t timeZone;
     bool soundEnable;
     int16_t screenBrightness;
