+++ Software/X-Track/USER/App/Pages/SystemInfos/SystemInfosView.cpp
@@ -545,8 +545,8 @@ void SystemInfosView::SetSport(
 }
 
 void SystemInfosView::SetGPS(
-    float lat,
-    float lng,
+    double lat,
+    double lng,
     float alt,
     const char* utc,
     float course,
